backend/ai_recruitment/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
python-multipart>=0.0.9
sqlalchemy>=2.0.32
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
langchain-text-splitters>=0.3.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
pypdf>=5.0.0
python-docx>=1.1.2
streamlit>=1.39.0
requests>=2.32.0
